% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smn1.R
\name{smn1}
\alias{smn1}
\title{A function to handle the detection of mutation in SMN1 gene}
\usage{
smn1(group = c("sudo", "docker"), scratch.folder, data.folder, threads = 4)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{threads, }{number of processors to be used in the analysis}
}
\description{
This function executes a docker that produces as output for the PG SMN1 pipeline
}
\examples{
\dontrun{
    #running smn1
    smn1(group="docker", scratch.folder="/data/scratch", data.folder=getwd(), threads=4)
}

}
\author{
Raffaele Calogero,raffaele.calogero [at] unito [dot] com, University of Torino
}
